vit-similarity-model/requirements.txt
numpy==2.1.2
torch==2.4.1
torch-model-archiver==0.9.0
torchvision== 0.19.1
pillow==10.4.0
pyyaml==6.0.2
pandas==2.2.2
scikit-learn==1.5.0
asyncio==3.4.3
aiohttp==3.10.9
sqlalchemy==2.0.36
pymysql==1.1.1
cloud-sql-python-connector==1.13.0
google-cloud-storage==2.18.2
matplotlib==3.9.2
opencv-python==4.10.0.84
